fix: fix some tests
