+++ jest-setup.ts
@@ -1,3 +1,6 @@
 import '@testing-library/jest-dom';
 import '@testing-library/dom';
 import '@testing-library/react';
+import fetchMock from 'jest-fetch-mock';
+
+fetchMock.enableMocks();
