+++ src/__mocks__/server.ts
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node';
+import { peopleHandlers } from './handlers';
+
+export const server = setupServer(...peopleHandlers);
